Action Framework - Smoothed Out Leap Drastically.
Z Root Motion only works in Flying Mode.
Created Action Framework to sync up player logic with animation logic.
CustomBlueprintEvents via Animation States only works with AnimNotify_ prefix
ActionManager tracks current action and can determine if transitions are allowed.
Root Motion of Animation Sequences only work if you open the AnimBP, close it, then reopen it, secret option to enable it for non montages.
Reanimated leap as I needed to animate it on the spot, and handle the movement in engine.
Tinkerred with the root motion of the leap until I was happy.